+++ client/utils.go
@@ -26,8 +26,8 @@ func (a Address) EthAddress() common.Address {
 const (
 	// CompoundBAT is the address of the cBAT contract
 	CompoundBAT = Address("0x6c8c6b02e7b2be14d4fa6022dfd6d75921d90e4e")
-	// CompoundDAI is the address of the cDAI contract
-	CompoundDAI = Address("0xf5dce57282a584d2746faf1593d3121fcac444dc")
+	// CompoundSAI is the address of the cSAI contract
+	CompoundSAI = Address("0xf5dce57282a584d2746faf1593d3121fcac444dc")
 	// CompoundETH is the address of the cETH contract
 	CompoundETH = Address("0x4ddc2d193948926d02f9b1fe9e1daa0718270ed5")
 	// CompoundREP is the address of the cREP contract
@@ -48,7 +48,7 @@ var (
 	// CompoundTokens is map containing the name, and address of all compound tokens
 	CompoundTokens = map[string]Address{
 		"cBAT":  CompoundBAT,
-		"cDAI":  CompoundDAI,
+		"cSAI":  CompoundSAI,
 		"cETH":  CompoundETH,
 		"cREP":  CompoundREP,
 		"cUSDC": CompoundUSDC,
